Allow override of mounts file
When in debug mode the default /proc/mounts source can be overridden to customize the source of fanotify events.

The intent here is to restrict events to specific locations when testing to make interpreting the event stream easier.

The alternative mounts file will expect the same format as /proc/mounts, allowing us to head, tail, or grep sublists from /proc/mounts into a new file that fapolicyd will use as the mount source.